+++ kadu-core/core/core.cpp
@@ -495,22 +495,22 @@ void Core::deleteOldConfigurationFiles()
 	QDir oldConfigs(KaduPaths::instance()->profilePath(), "kadu-0.12.conf.xml.backup.*", QDir::Name, QDir::Files);
 	if (oldConfigs.count() > 20)
 		for (unsigned int i = 0, max = oldConfigs.count() - 20; i < max; ++i)
-			QFile::remove(KaduPaths::instance()->profilePath() + oldConfigs[i]);
+			QFile::remove(KaduPaths::instance()->profilePath() + oldConfigs[static_cast<int>(i)]);
 
 	QDir oldConfigs2(KaduPaths::instance()->profilePath(), "kadu-0.6.6.conf.xml.backup.*", QDir::Name, QDir::Files);
 	if (oldConfigs2.count() > 20)
 		for (unsigned int i = 0, max = oldConfigs2.count() - 20; i < max; ++i)
-			QFile::remove(KaduPaths::instance()->profilePath() + oldConfigs2[i]);
+			QFile::remove(KaduPaths::instance()->profilePath() + oldConfigs2[static_cast<int>(i)]);
 
 	QDir oldBacktraces(KaduPaths::instance()->profilePath(), "kadu.backtrace.*", QDir::Name, QDir::Files);
 	if (oldBacktraces.count() > 20)
 		for (unsigned int i = 0, max = oldBacktraces.count() - 20; i < max; ++i)
-			QFile::remove(KaduPaths::instance()->profilePath() + oldBacktraces[i]);
+			QFile::remove(KaduPaths::instance()->profilePath() + oldBacktraces[static_cast<int>(i)]);
 
 	QDir oldDebugs(KaduPaths::instance()->profilePath(), "kadu.log.*", QDir::Name, QDir::Files);
 	if (oldDebugs.count() > 20)
 		for (unsigned int i = 0, max = oldDebugs.count() - 20; i < max; ++i)
-			QFile::remove(KaduPaths::instance()->profilePath() + oldDebugs[i]);
+			QFile::remove(KaduPaths::instance()->profilePath() + oldDebugs[static_cast<int>(i)]);
 
 	kdebugf2();
 }
